Implement basic support for Session open both local and remote
Adds some event handlers and implement processing of local open and to
emit a Begin and handling of remote Begin to either open the locally
opened session or create a new Session based on remote open.

+++ qpid-proton4j-engine/src/main/java/org/apache/qpid/proton4j/engine/EngineHandler.java
@@ -82,7 +82,7 @@ public interface EngineHandler {
         context.fireWrite(header);
     }
 
-    default void handleWrite(EngineHandlerContext context, Performative performative, short channel, ProtonBuffer payload, Runnable payloadToLarge) {
+    default void handleWrite(EngineHandlerContext context, Performative performative, int channel, ProtonBuffer payload, Runnable payloadToLarge) {
         context.fireWrite(performative, channel, payload, payloadToLarge);
     }
